requirements.txt
PyJWT~=2.8.0
pytz~=2023.3.post1
DBUtils~=3.0.3
PyMySQL~=1.1.0
Flask~=2.2.5
loguru~=0.7.2